% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_function.R
\name{apply_function}
\alias{apply_function}
\title{apply_functionUI}
\usage{
apply_function(input, output, session, variable = NULL, name = "")
}
\arguments{
\item{input}{Not a real parameter, should not be set manually. Done by callModule automatically.}

\item{output}{Not a real parameter, should not be set manually. Done by callModule automatically.}

\item{session}{Not a real parameter, should not be set manually. Done by callModule automatically.}

\item{variable}{Numeric. Vector containing dataset to apply function on.}

\item{name}{Character. Name of the variable (used only on UI part).}
}
\value{
Server logic
}
\description{
This function has to be set in the Server part of a shiny application
    It add a windows containing a radioButton to apply a function
    on a numeric vector.
}
\examples{
\dontrun{
# In UI :
apply_functionUI(id = "mod2")
# In Server
data_module2   <- callModule(module = apply_function,
                             id = "mod2",
                             variable = reactive(dataset$data_var_x),
                             name = "Axis X")
}
}
