% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_function.R
\name{apply_functionUI}
\alias{apply_functionUI}
\title{apply_functionUI}
\usage{
apply_functionUI(id)
}
\arguments{
\item{id}{An id that will be used to create a namespace}
}
\value{
UI page
}
\description{
This function has to be set in the UI part of a shiny application
    It add a windows containing a radioButton to apply a function
    on a numeric vector.
apply_function function has to be set in the Server part.
}
\examples{
\dontrun{
# In UI :
apply_functionUI(id = "mod2")
# In Server
data_module2   <- callModule(module = apply_function,
                             id = "mod2",
                             variable = reactive(dataset$data_var_x))
}
}
