% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\name{data_info}
\alias{data_info}
\title{data_info}
\usage{
data_info(input, output, session, data, data_name)
}
\arguments{
\item{input, output, session}{Not real parameter, should not be set manually. Done by callModule automatically.}

\item{data}{data.frame (not a reactive)}

\item{data_name}{Dataset name (not a reactive)}
}
\value{
Server logic
}
\description{
This function has to be set in the Server part of a shiny application.
load_dataUI function has to be set in the UI part.
}
\examples{
\dontrun{
# In UI :
data_infoUI(id = "mod1")
# In Server
callModule(module = data_info, id = "mod1")
}
}
