% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\name{data_infoUI}
\alias{data_infoUI}
\title{data_infoUI}
\usage{
data_infoUI(id)
}
\arguments{
\item{id}{An id that will be used to create a namespace}
}
\value{
UI page
}
\description{
This function has to be set in the UI part of a shiny application
it adds a panel of width 4. Its title is the dataset name and its content is the
dataset dimensions.
data_info function has to be set in the Server part.
}
\examples{
\dontrun{
# In UI :
data_infoUI(id = "mod1")
# In Server
callModule(module = data_info, id = "mod1")
}
}
