% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funHistory.R
\name{funHistoryUI}
\alias{funHistoryUI}
\title{apply_functionUI}
\usage{
funHistoryUI(id)
}
\arguments{
\item{id}{An id that will be used to create a namespace}
}
\value{
UI page
}
\description{
This function has to be set in the UI part of a shiny application
    It add a windows containing a radioButton to apply a function
    on a numeric vector.
apply_function function has to be set in the Server part.
}
\examples{
\dontrun{
# In UI
funHistoryUI(id = "id")

# In Server
callModule(module = funHistory, id = "id", histo = reactive(histo$fun_applied_x), name = "X vector")
}
}
