% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{load_dataUI}
\usage{
load_data(input, output, session)
}
\arguments{
\item{input}{Not a real parameter, should not be set manually. Done by callModule automatically.}

\item{output}{Not a real parameter, should not be set manually. Done by callModule automatically.}

\item{session}{Not a real parameter, should not be set manually. Done by callModule automatically.}
}
\value{
Server logic
}
\description{
This function has to be set in the Server part of a shiny application
    it adds the load data windows.
    load_dataUI function has to be set in the UI part.
}
\examples{
\dontrun{
# In UI :
load_dataUI(id = "mod1")
# In Server
data_module1 <- callModule(module = load_data, id = "mod1")
}
}
