% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dataset.R
\name{load_dataset}
\alias{load_dataset}
\title{load_datasetUI}
\usage{
load_dataset(input, output, session)
}
\arguments{
\item{input}{Not a real parameter, should not be set manually. Done by callModule automatically.}

\item{output}{Not a real parameter, should not be set manually. Done by callModule automatically.}

\item{session}{Not a real parameter, should not be set manually. Done by callModule automatically.}
}
\value{
Server logic
}
\description{
This function has to be set in the Server part of a shiny application
    it adds the load data windows.
    load_datasetUI function has to be set in the UI part.
}
\examples{
\dontrun{
# In UI :
load_datasetUI(id = "mod1")
# In Server
data_module1 <- callModule(module = load_dataset, id = "mod1")
}
}
