% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dataset.R
\name{load_datasetUI}
\alias{load_datasetUI}
\title{load_datasetUI}
\usage{
load_datasetUI(id)
}
\arguments{
\item{id}{An id that will be used to create a namespace}
}
\value{
UI page
}
\description{
This function has to be set in the UI part of a shiny application
it adds the load data box.
load_dataset function has to be set in the Server part.
}
\examples{
\dontrun{
# In UI :
load_datasetUI(id = "mod1")
# In Server
data_module1 <- callModule(module = load_dataset, id = "mod1")
}
}
