% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_dataUI}
\alias{load_dataUI}
\title{load_dataUI}
\usage{
load_dataUI(id)
}
\arguments{
\item{id}{An id that will be used to create a namespace}
}
\value{
UI page
}
\description{
This function has to be set in the UI part of a shiny application
it adds the load data box.
load_data function has to be set in the Server part.
}
\examples{
\dontrun{
# In UI :
load_dataUI(id = "mod1")
# In Server
data_module1 <- callModule(module = load_data, id = "mod1")
}
}
