% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadModulesCSS}
\alias{loadModulesCSS}
\title{loadModulesCSS}
\usage{
loadModulesCSS(cssFilenames = c())
}
\arguments{
\item{cssFilenames}{Character css file to include (extension mandatory) Default c() means all.}
}
\value{
Load css
}
\description{
This function has to be set in the ui part to link CSS file.
}
\examples{
\dontrun{
# In UI
shinyModulesTuto::loadModulesCSS()
}
}
