% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_modules.R
\name{merge_modules}
\alias{merge_modules}
\title{merge_modules}
\usage{
merge_modules(input, output, session, data, name)
}
\arguments{
\item{input, output, session}{mandatory arguments for modules to be valid. These
should not to be defined as they will be handled by shiny.}

\item{data}{Numeric vector (not a reactive variable).}

\item{name}{Variable name (not a reactive variable).}
}
\value{
reactiveValues with 2 slots :
name : name of the variable.
nb_funs : number of functions applied on vector.
}
\description{
This function has to be set in the Server part of a shiny application.
    The module merges modules apply_function, apply_scale, show_data & funHistory
    merge_modulesUI function has to be set in the UI part.
}
\examples{
\dontrun{
# In UI :
merge_modulesUI(id = "id1")
# In Server
data_module1 <- callModule(
  module = merge_modules, 
  id = "id1", 
  data = iris$Sepal.Length, 
  name = "Sepal.Length")
}
}
