% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_modules.R
\name{merge_modulesUI}
\alias{merge_modulesUI}
\title{merge_modulesUI}
\usage{
merge_modulesUI(id)
}
\arguments{
\item{id}{namespace identifier for the module}
}
\value{
div
}
\description{
This function has to be set in the UI part of a shiny application.
    The module merges modules apply_function, apply_scale, show_data & funHistory
    Function merge_modules must be set in server part.
}
\examples{
\dontrun{
# In UI :
merge_modulesUI(id = "id1")
# In Server
data_module1 <- callModule(
  module = merge_modules, 
  id = "id1", 
  data = iris$Sepal.Length, 
  name = "Sepal.Length")
}
}
