% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useExamples.R
\name{runEx}
\alias{runEx}
\title{Run Shiny application}
\usage{
runEx(app, showcase = TRUE)
}
\arguments{
\item{app}{Character Application folder name}

\item{showcase}{Logical if app should be launched in showcase mode}
}
\description{
Launch shiny applications present in shinyModulesTuto/applications/
}
\examples{
\dontrun{
runEx(app = "example1")
}
}
