% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_data.R
\name{show_data}
\alias{show_data}
\title{show_data}
\usage{
show_data(input, output, session, variable = NULL,
  variable_name = NULL, useggplot = FALSE)
}
\arguments{
\item{input}{Not a real parameter, should not be set manually. Done by callModule automatically.}

\item{output}{Not a real parameter, should not be set manually. Done by callModule automatically.}

\item{session}{Not a real parameter, should not be set manually. Done by callModule automatically.}

\item{variable}{Reactive numeric vector}

\item{variable_name}{Reactive character}

\item{useggplot}{Boolean if use ggplot or basic hist function}
}
\value{
Server logic
}
\description{
This function has to be set in the Server part of a shiny application
    it adds the load data windows.
    show_dataUI function has to be set in the UI part.
}
\examples{
\dontrun{
# In UI :
show_dataUI(id = "mod1")
# In Server
data_module1 <- callModule(module = show_data, id = "mod3", variable = variable, variable_name = variable_name)
}
}
