% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funHistory.R
\name{funHistory}
\alias{funHistory}
\title{funHistory}
\usage{
funHistory(input, output, session, histo = NULL)
}
\arguments{
\item{input}{Not a real parameter, should not be set manually. Done by callModule automatically.}

\item{output}{Not a real parameter, should not be set manually. Done by callModule automatically.}

\item{session}{Not a real parameter, should not be set manually. Done by callModule automatically.}

\item{histo}{Numeric. Vector containing function applied.}
}
\value{
Server logic
}
\description{
This function has to be set in the Server part of a shiny application
    It add a windows containing a radioButton to apply a function
    on a numeric vector.
}
\examples{
\dontrun{
# In UI
funHistoryUI(id = "id")

# In Server
callModule(module = funHistory, id = "id", histo = reactive(rv$fun_historic))
}
}
